# openapi.yaml
openapi: 3.0.1
info:
  title: TODO Plugin
  description: A plugin that allows the user to create and manage a TODO list using ChatGPT.
  version: 'v1'
servers:
  - url: https://greasy-jessika-digital-alpha.koyeb.app/
paths:
  /get_investor_cik:
    post:
      operationId: get_investor_cik
      summary: Get an array of Investor name in database and it's CIK from it's regular name
      description: Something
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetInvestorCikRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetInvestorCikResponse'
      responses:
        "300":
          description: CIK NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetInvestorCikResponseIfNull'
  /get_issuer_cusip:
    post:
      operationId: get_issuer_cusip
      summary: Get the cusip of a particular asset/issuer from it's name
      description: something
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetIssuerCusipRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetIssuerCusipResponse'
      responses:
        "300":
          description: CUSIP NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetIssuerCusipResponseIfNull'
  /get_investor_name:
    post:
      operationId: get_investor_name
      summary: Get the Investor name for a CIK
      description: something
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetInvestorNameRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetInvestorNameResponse'
      responses:
        "300":
          description: INVESTOR NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetInvestorNameResponseIfNull'
  /get_filings:
    post:
      operationId: get_filings
      summary: Query of a database to get records on a Db having 13F documents
      description: run a query in database having 13f documents with the schema \n \ninvestor_cik,\nperiod_of_report timestamp,\n"TITLE OF CLASS",\n"CUSIP",\n"FIGI",\n"VALUE (to the nearest dollar)",\n"SHRS OR PRN AMT",\n"SH/PRN" and table name aggregate_filings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetFilingsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFilingsResponse'
      responses:
        "300":
          description: QUERY NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFilingsResponseIfNull'
components:
  schemas:
    GetInvestorCikRequest:
      title: GetInvestorCikRequest
      required:
        - investor_name
      type: object
      properties:
        investor_name:
          title: InvestorName
          type: string
          description: The Name of investor we want to know CIK.
    GetIssuerCusipRequest:
      title: GetIssuerCusipRequest
      required:
        - issuer_name
      type: object
      properties:
        issuer_name:
          title: IssuerName
          type: string
          description: The Name of issuer/asset we want to know cusip.
    GetInvestorNameRequest:
      title: GetInvestorNameRequest
      required:
        - investor_cik
      type: object
      properties:
        investor_cik:
          title: InvestorCik
          type: string
          description: The CIK of investor for knowing the name.
    GetFilingsRequest:
      title: GetFilingsRequest
      required:
        - db_query
      type: object
      properties:
        db_query:
          title: Database Query
          type: string
          description: The Query that has to be ran on Db.
    GetInvestorCikResponse:
      title: GetInvestorCikResponse
      required:
        - results
      type: object
      properties:
        results:
          title: Results
          type: object
          description: Object containing investor information
          properties:
            investor_name:
              title: Investor Name
              type: array
              description: Array of investor names
              items:
                type: string
            cik:
              title: CIK
              type: array
              description: Array of CIK (Central Index Key) numbers
              items:
                type: string
    GetInvestorCikResponseIfNull:
      title: GetInvestorCikResponseIfNull
      required:
        - results
      type: object
      properties:
        results:
          title: Results
          type: string
          description: If the specified investor is not found
    GetIssuerCusipResponse:
      title: GetIssuerCusipResponse
      required:
        - results
      type: object
      properties:
        results:
          title: Results
          type: object
          description: Object containing investor information
          properties:
            name_of_issuer:
              title: Issuer/Asset Name
              type: array
              description: Array of investor/assets names
              items:
                type: string
            cusip:
              title: CUSIP
              type: array
              description: Array of CUSIPs
              items:
                type: string
            issuer:
              title: First 6 characters of the cusip
              type: array
              description: Array of first 6 characters of the cusip this should be used while searching issuers
              items:
                type: string
    GetIssuerCusipResponseIfNull:
      title: GetIssuerCusipResponseIfNull
      required:
        - results
      type: object
      properties:
        results:
          title: Results
          type: string
          description: If the specified asset/issuer is not found
    GetInvestorNameResponse:
      title: GetInvestorNameResponse
      required:
        - results
      type: object
      properties:
        results:
          title: Results
          type: string
          description: Name of the investor
    GetInvestorNameResponseIfNull:
      title: GetInvestorNameResponseIfNull
      required:
        - results
      type: object
      properties:
        results:
          title: Results
          type: string
          description: If the specified CIK is not found
    GetFilingsResponse:
      title: GetFilingsResponse
      required:
        - results
      type: object
      properties:
        results:
          title: Results
          type: string
          description: CSV of values from the database
    GetFilingsResponseIfNull:
      title: GetFilingsResponseIfNull
      required:
        - results
      type: object
      properties:
        results:
          title: Results
          type: string
          description: If the specified query is not found